alteracao int-float
como se trata de notas de 0 a 10, talvez seja melhor usar float no lugar de int para digitar as notas, haja vista que elas podem receber valores decimais

+++ 04.py
@@ -8,9 +8,9 @@ def calcular_media(n1, n2, n3):
         print(f'Reprovado, a média é , {media:.2f}')
     return media
 if __name__ == '__main__':
-    n1= int(input('Digite a primeira nota: '))
-    n2= int(input('digite a segunda nota: '))
-    n3= int(input('digite a terceira nota: '))
+    n1= float(input('Digite a primeira nota: '))
+    n2= float(input('digite a segunda nota: '))
+    n3= float(input('digite a terceira nota: '))
     calcular_media(n1, n2, n3)
